# Add benchmark and character conformance tests