--- a/.travis.yml
+++ b/.travis.yml
@@ -12,8 +12,12 @@ matrix:
 notifications:
   webhooks: http://build.servo.org:54856/travis
 
+before_script:
+    - pip install git+https://github.com/euclio/travis-cargo@kcov --user && export PATH=$HOME/.local/bin:$PATH
+
 script:
   - cargo build --verbose
-  - cargo test --verbose
-  - cargo test --verbose --features with_serde
-  - cargo doc --verbose --no-deps
+  - cargo test  --verbose
+  - cargo test  --verbose --features with_serde
+  - cargo doc   --verbose --no-deps
+  - travis-cargo bench