--- a/Cargo.toml
+++ b/Cargo.toml
@@ -21,4 +21,5 @@ serde_derive = {version = ">=0.8, <2.0", optional = true}
 
 [features]
 default = []
+unstable = []  # Use in benches/
 with_serde = ["serde", "serde_test", "serde_derive"]