--- a/benches/udhr_data/README.md
+++ b/benches/udhr_data/README.md
@@ -1,0 +1,3 @@
+# A collection of UDHR in Unicode text files
+
+Source: http://unicode.org/udhr/