… benches/udhr_data/bidi/udhr_aii.txt
… benches/udhr_data/bidi/udhr_arb.txt
… benches/udhr_data/bidi/udhr_eng.txt
… benches/udhr_data/bidi/udhr_heb.txt
… benches/udhr_data/bidi/udhr_mly_arab.txt
… benches/udhr_data/bidi/udhr_mly_latn.txt
… benches/udhr_data/bidi/udhr_pes_1.txt
… benches/udhr_data/bidi/udhr_pes_2.txt
… benches/udhr_data/bidi/udhr_pnb.txt
… benches/udhr_data/bidi/udhr_skr.txt
… benches/udhr_data/bidi/udhr_uig_arab.txt
… benches/udhr_data/bidi/udhr_uig_latn.txt
… benches/udhr_data/bidi/udhr_urd.txt
… benches/udhr_data/bidi/udhr_urd_2.txt
… benches/udhr_data/bidi/udhr_ydd.txt
… benches/udhr_data/ltr/udhr_acu_1.txt
… benches/udhr_data/ltr/udhr_auc.txt
… benches/udhr_data/ltr/udhr_knc.txt
… benches/udhr_data/ltr/udhr_krl.txt
… benches/udhr_data/ltr/udhr_kwi.txt
… benches/udhr_data/ltr/udhr_lot.txt
… benches/udhr_data/ltr/udhr_piu.txt
… benches/udhr_data/ltr/udhr_qug.txt
… benches/udhr_data/ltr/udhr_snn.txt
… benches/udhr_data/ltr/udhr_tiv.txt