## Add benchmark tests using UDHR text files

The bench functions are put in two groups: `for_ltr_texts` and
`for_bidi_texts`. This will allow us measure improvements for both
all-LTR and bidirectional situations.

Current results:
```
test bench_bidi_info_new_and_reordered_for_bidi_texts ... bench: 9,711,485 ns/iter (+/- 917,086)
test bench_bidi_info_new_and_reordered_for_ltr_texts  ... bench: 6,496,454 ns/iter (+/- 829,583)
test bench_bidi_info_new_for_bidi_texts               ... bench: 9,813,185 ns/iter (+/- 1,180,446)
test bench_bidi_info_new_for_ltr_texts                ... bench: 6,811,823 ns/iter (+/- 1,079,303)
```

## [tests] Add test_character_conformance() and some cleanup

* Add `test_character_conformance()`, which runs against UCD's
`BidiCharacterTest.txt` test file.

* Clean up and add line_num and actual values to fail output data.

<!-- Reviewable:start -->
---
This change is [<img src="https://reviewable.io/review_button.svg" height="34" align="absmiddle" alt="Reviewable"/>](https://reviewable.io/reviews/servo/unicode-bidi/33)
<!-- Reviewable:end -->