@behnam 2017-05-16
Finally got the nightly bench working on travis! The trick is that have to create an `unstable` branch, to be used by `travis-cargo`, and it's an error to not have one in `Cargo.toml`.

```
test bench_bidi_info_new_and_reordered_for_bidi_texts ... bench:  12,569,450 ns/iter (+/- 913,045)
test bench_bidi_info_new_and_reordered_for_ltr_texts  ... bench:   8,172,522 ns/iter (+/- 81,493)
test bench_bidi_info_new_for_bidi_texts               ... bench:  12,488,570 ns/iter (+/- 238,099)
test bench_bidi_info_new_for_ltr_texts                ... bench:   8,339,091 ns/iter (+/- 780,304)
```