@behnam 2017-05-17
## [tests] Add test_character_conformance() and some cleanup

* Add `test_character_conformance()`, which runs against UCD's
`BidiCharacterTest.txt` test file.

* Clean up and add line_num and actual values to fail output data.