@bors-servo 2017-05-18
:pushpin: Commit 54cb4d9 has been approved by `mbrubeck`

<!-- @bors-servo r=mbrubeck 54cb4d932fa13889e7b07c7088373007af3f0e70 -->