@bors-servo 2017-05-18
:hourglass: Testing commit 54cb4d932fa13889e7b07c7088373007af3f0e70 with merge 019b6712e5c837aadb3fb97ebfe2eb75a0a9e5c1...