@bors-servo 2017-05-18
:sunny: Test successful - [status-travis](https://travis-ci.org/servo/unicode-bidi/builds/233839729?utm_source=github_status&utm_medium=notification)
Approved by: mbrubeck
Pushing 019b6712e5c837aadb3fb97ebfe2eb75a0a9e5c1 to master...